\name{simpleboot-internal}
\alias{simpleboot-internal}
\alias{lm.boot.resample}
\alias{lo.boot.resample}
\alias{model.frame.lm.simpleboot}

\title{Internal simpleboot functions}
\description{
  Internal \code{simpleboot} functions.
}
\usage{
lm.boot.resample(lm.object, R, rows, new.xpts, weights)
lo.boot.resample(lo.object, R, rows, new.xpts, weights)
model.frame.lm.simpleboot(formula, ...)
}
\details{
  This functions are not to be called by the user.
}
\keyword{internal}

\author{Roger D. Peng}
