\name{uclaboot-internal}
\alias{uclaboot-internal}
\alias{lm.boot.resample}
\alias{lo.boot.resample}
\alias{model.frame.lm.uclaboot}

\title{Internal uclaboot functions}
\description{
  Internal \code{uclaboot} functions.
}
\usage{
lm.boot.resample(lm.object, R, rows, new.xpts, weights)
lo.boot.resample(lo.object, R, rows, new.xpts, weights)
model.frame.lm.uclaboot(formula, ...)
}
\details{
  This functions are not to be called by the user.
}
\keyword{internal}

\author{Roger D. Peng}
